+++ classes/output/core_renderer.php
@@ -282,4 +282,76 @@ class core_renderer extends \theme_boost\output\core_renderer {
 
         return ' id="'. $this->body_id().'" class="'.$this->body_css_classes($additionalclasses).'"';
     }
+
+    /**
+     * Wrapper for header elements.
+     *
+     * This renderer function is copied and modified from /lib/outputrenderers.php
+     *
+     * @return string HTML to display the main header.
+     */
+    public function full_header() {
+        $pagetype = $this->page->pagetype;
+        $homepage = get_home_page();
+        $homepagetype = null;
+        // Add a special case since /my/courses is a part of the /my subsystem.
+        if ($homepage == HOMEPAGE_MY || $homepage == HOMEPAGE_MYCOURSES) {
+            $homepagetype = 'my-index';
+        } else if ($homepage == HOMEPAGE_SITE) {
+            $homepagetype = 'site-index';
+        }
+        if ($this->page->include_region_main_settings_in_header_actions() &&
+                !$this->page->blocks->is_block_present('settings')) {
+            // Only include the region main settings if the page has requested it and it doesn't already have
+            // the settings block on it. The region main settings are included in the settings block and
+            // duplicating the content causes behat failures.
+            $this->page->add_header_action(html_writer::div(
+                    $this->region_main_settings_menu(),
+                    'd-print-none',
+                    ['id' => 'region-main-settings-menu']
+            ));
+        }
+
+        $header = new \stdClass();
+        $header->settingsmenu = $this->context_header_settings_menu();
+        $header->contextheader = $this->context_header();
+        $header->hasnavbar = empty($this->page->layout_options['nonavbar']);
+        $header->navbar = $this->navbar();
+        $header->pageheadingbutton = $this->page_heading_button();
+        $header->courseheader = $this->course_header();
+        $header->headeractions = $this->page->get_header_actions();
+
+        // Add the course header image for rendering.
+        if ($this->page->pagelayout == 'course' && (get_config('theme_boost_union', 'courseheaderimageenabled')
+                        == THEME_BOOST_UNION_SETTING_SELECT_YES)) {
+            // If course header images are activated, we get the course header image url
+            // (which might be the fallback image depending on the course settings and theme settings).
+            $header->courseheaderimageurl = theme_boost_union_get_course_header_image_url();
+            // Additionally, get the course header image height.
+            $header->courseheaderimageheight = get_config('theme_boost_union', 'courseheaderimageheight');
+            // Additionally, get the course header image position.
+            $header->courseheaderimageposition = get_config('theme_boost_union', 'courseheaderimageposition');
+            // Additionally, get the template context attributes for the course header image layout.
+            $courseheaderimagelayout = get_config('theme_boost_union', 'courseheaderimagelayout');
+            switch($courseheaderimagelayout) {
+                case THEME_BOOST_UNION_SETTING_COURSEIMAGELAYOUT_HEADINGABOVE:
+                    $header->courseheaderimagelayoutheadingabove = true;
+                    $header->courseheaderimagelayoutstackedclass = '';
+                    break;
+                case THEME_BOOST_UNION_SETTING_COURSEIMAGELAYOUT_STACKEDDARK:
+                    $header->courseheaderimagelayoutheadingabove = false;
+                    $header->courseheaderimagelayoutstackedclass = 'dark';
+                    break;
+                case THEME_BOOST_UNION_SETTING_COURSEIMAGELAYOUT_STACKEDLIGHT:
+                    $header->courseheaderimagelayoutheadingabove = false;
+                    $header->courseheaderimagelayoutstackedclass = 'light';
+                    break;
+            }
+        }
+
+        if (!empty($pagetype) && !empty($homepagetype) && $pagetype == $homepagetype) {
+            $header->welcomemessage = \core_user::welcome_message();
+        }
+        return $this->render_from_template('core/full_header', $header);
+    }
 }